requirement.txt
gitpython==3.1.37
langchain
datasets==2.14.4
torch==2.0.1
transformers==4.32.1
bitsandbytes==0.41.1
trl==0.7.1
accelerate
peft==0.5.0
langchain
esprima
fsspec
